examples/model_serving/requirements.txt
python-dotenv
uvicorn
pydantic
sentence_transformers
xformers
torch
torchvision
fastapi